[42] Move common methods in test to utility module.
