+++ test/baselines/semi/test_syntactic_filtering.py
@@ -1,15 +1,8 @@
-from typing import List, Tuple, Union
-from pathlib import Path
+from typing import List
 from unittest import TestCase
 
 from veniq.baselines.semi._syntactic_filter import syntactic_filter
-from veniq.baselines.semi._common_types import (
-    ExtractionOpportunity,
-    Statement as ExtractionStatement,
-)
-from veniq.ast_framework.block_statement_graph import build_block_statement_graph, Block, Statement
-from veniq.ast_framework import AST, ASTNodeType
-from veniq.utils.ast_builder import build_ast
+from .utils import get_method_ast, create_extraction_opportunity
 
 
 class FilteringTestCase(TestCase):
@@ -32,56 +25,14 @@ class FilteringTestCase(TestCase):
         self._opportunity_test_helper([5, 6, 7, 8, 9, 11, 12, 15, 19, 20, 21, 22], True)
 
     def _opportunity_test_helper(
-        self, extraction_opportunity_statements_ids: List[int], is_opportunity_correct: bool
+        self, extraction_opportunity_statements_lines: List[int], is_opportunity_correct: bool
     ):
-        method_ast = self._get_method_ast()
-        extraction_opportunity, block_statement_graph = self._create_extraction_opportunity(
-            method_ast, extraction_opportunity_statements_ids
+        method_ast = get_method_ast("SyntacticFilterTest.java", "Test", "testMethod")
+        extraction_opportunity, block_statement_graph = create_extraction_opportunity(
+            method_ast, extraction_opportunity_statements_lines
         )
+
         self.assertEqual(
-            self._is_correct_extraction_opportunity(extraction_opportunity, block_statement_graph),
+            syntactic_filter(extraction_opportunity, block_statement_graph),
             is_opportunity_correct,
         )
-
-    @staticmethod
-    def _create_extraction_opportunity(
-        method_ast: AST, statements_lines: List[int]
-    ) -> Tuple[ExtractionOpportunity, Block]:
-        extraction_opportunity_list: List[ExtractionStatement] = []
-        block_statement_graph = build_block_statement_graph(method_ast)
-
-        def fill_extraction_opportunity(node: Union[Block, Statement]):
-            nonlocal extraction_opportunity_list
-            if isinstance(node, Statement) and node.node.line in statements_lines:
-                extraction_opportunity_list.append(node.node)
-
-        block_statement_graph.traverse(fill_extraction_opportunity)
-        return tuple(extraction_opportunity_list), block_statement_graph
-
-    @staticmethod
-    def _is_correct_extraction_opportunity(
-        extraction_opportunity: ExtractionOpportunity,
-        block_statement_graph: Block,
-    ) -> bool:
-        return syntactic_filter(extraction_opportunity, block_statement_graph)
-
-    @staticmethod
-    def _get_method_ast() -> AST:
-        current_directory = Path(__file__).absolute().parent
-        filepath = current_directory / "SynctacticFilterTest.java"
-        ast = AST.build_from_javalang(build_ast(str(filepath)))
-
-        try:
-            class_declaration = next(
-                node
-                for node in ast.get_root().types
-                if node.node_type == ASTNodeType.CLASS_DECLARATION and node.name == "Example"
-            )
-
-            method_declaration = next(
-                node for node in class_declaration.methods if node.name == "exampleMethod"
-            )
-        except StopIteration:
-            raise RuntimeError(f"Failed to find method 'exampleMethod' in class 'Example' in file {filepath}")
-
-        return ast.get_subtree(method_declaration)
